requirements.txt
anyio==3.6.2
bcrypt==4.0.1
bson==0.5.10
click==8.1.3
ecdsa==0.18.0
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
loguru==0.7.0
orjson==3.8.10
passlib==1.7.4
pyasn1==0.5.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
